Implemented MapSPI methods to make tower module unloadable
There seemed to be something wrong with the collision-method in MapPlugin. Therefore I re-wrote the getTilesEntityIsOn-method. I kept the old methods as comments if anyone needs them for something.
I also added a list of entities to the checkIfTileIsOccupied-method. This enables us to pass entities the method should ignore when checking for occupants.